+++ src/App.jsx
@@ -1,18 +1,12 @@
+import ToDoList from './ToDoList.jsx';
+import TodoForm from './TodoForm.jsx';
 
 function App() {
-  const todos = [
-    { id: 1, title: 'review resources' },
-    { id: 2, title: 'take notes' },
-    { id: 3, title: 'code out app' },
-  ];
   return (
     <div>
-      <h1>My Todos</h1>
-      <ul>
-        {todos.map((todo) => (
-          <li key={todo.id}>{todo.title}</li>
-        ))}
-      </ul>
+      <h1>Todos list</h1>
+      <TodoForm />
+      <ToDoList />
     </div>
   );
 }
